+++ src/interpolation.lua
@@ -4,6 +4,7 @@
 
 
 --- @meta
+--- this module implement some default interpolation methods for the api.
 
 
 --- -----------------------------------------------------------
@@ -16,7 +17,7 @@
 --- 	interpolation.linear(0, 8, 0.5) -- 4
 --- 	interpolation.quadratic_in(0, 8, 0.5) -- 2
 --- 	interpolation.quadratic_out(0, 8, 0.5) -- 4.5
---- @type { [string]: fun(x: number, y: number, t: number) }
+--- @type table<string, fun(x: number, y: number, t: number)>
 BeTweenApi.interpolation = {}